sports/requirements.txt
langchain>=0.0.350,<0.3.0
langchain-community>=0.2.9
langchain-core>=0.1.9
langchain-google-genai>=0.0.5
chromadb==0.4.18
google-generativeai>=0.3.2
google-ai-generativelanguage>=0.4.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
streamlit==1.29.0
uvicorn==0.24.0
pydantic>=1.9,<3.0.0
python-multipart==0.0.6
